Volume and compatibility fixes
Lowered resolution
Lowered volume on alarm
Created JApplet version
Created Windows version

+++ src/ld28/player/GoalObject.java
@@ -49,5 +49,5 @@ public class GoalObject extends Sprite {
 		}
 	}
 	
-	public void update(float dt) {}
+	public void update(double dt) {}
 }
